Move ExecutorThreadingServiceStats to metrics directory.
