+++ searchcore/src/vespa/searchcore/proton/server/CMakeLists.txt
@@ -34,7 +34,6 @@ vespa_add_library(searchcore_server STATIC
     documentsubdbcollection.cpp
     emptysearchview.cpp
     executor_thread_service.cpp
-    executor_threading_service_stats.cpp
     executorthreadingservice.cpp
     fast_access_doc_subdb.cpp
     fast_access_doc_subdb_configurer.cpp
