+++ searchcore/src/vespa/searchcore/proton/server/documentdb.cpp
@@ -11,7 +11,7 @@
 #include "maintenance_jobs_injector.h"
 #include "reconfig_params.h"
 #include "bootstrapconfig.h"
-#include "executor_threading_service_stats.h"
+#include <vespa/searchcore/proton/metrics/executor_threading_service_stats.h>
 #include <vespa/searchcore/proton/attribute/attribute_writer.h>
 #include <vespa/searchcore/proton/attribute/imported_attributes_repo.h>
 #include <vespa/searchcore/proton/common/eventlogger.h>
